Be more aggressive about meeting dev_count.
It is possible that when requesting multi_dest create_open, that we do
not have enough devices to suggest. To mitigate this, keep track of
devices we skipped due to being on the same host as other devices. Then
add these at the end to make up for the lacking devices.

If there's no way to get to dev_count, return what we have.